requirements.txt
openai
python-telegram-bot
tiktoken
beautifulsoup4
langchain
marvin
pandas
